#78 [feat] User의 테스트 코드를 리팩토링 한 값에 맞게 수정한다.
