+++ src/main/java/kusitms/backend/user/application/UserApplicationService.java
@@ -1,22 +1,21 @@
 package kusitms.backend.user.application;
 
 import jakarta.servlet.http.HttpServletResponse;
-import kusitms.backend.auth.dto.response.*;
-import kusitms.backend.user.infra.jwt.JWTUtil;
-import kusitms.backend.user.status.AuthErrorStatus;
 import kusitms.backend.global.exception.CustomException;
 import kusitms.backend.global.redis.DistributedLockManager;
 import kusitms.backend.global.redis.RedisManager;
 import kusitms.backend.global.util.CookieUtil;
-import kusitms.backend.user.application.dto.response.*;
-import kusitms.backend.user.domain.enums.ProviderStatusType;
-import kusitms.backend.user.domain.model.User;
-import kusitms.backend.user.domain.repository.UserRepository;
 import kusitms.backend.user.application.dto.request.CheckNicknameRequestDto;
 import kusitms.backend.user.application.dto.request.SendAuthCodeRequestDto;
 import kusitms.backend.user.application.dto.request.SignUpRequestDto;
 import kusitms.backend.user.application.dto.request.VerifyAuthCodeRequestDto;
+import kusitms.backend.user.application.dto.response.*;
+import kusitms.backend.user.domain.enums.ProviderStatusType;
+import kusitms.backend.user.domain.model.User;
+import kusitms.backend.user.domain.repository.UserRepository;
+import kusitms.backend.user.infra.jwt.JWTUtil;
 import kusitms.backend.user.infra.sms.SmsManager;
+import kusitms.backend.user.status.AuthErrorStatus;
 import kusitms.backend.user.status.UserErrorStatus;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
